refactor(mbk/leases): split lease_template_service.py (924 LOC) into focused modules
Splits a large multi-responsibility service into three focused modules,
matching the precedent set by PR #487 (signed_lease_service split):

- lease_template_service.py (830 LOC) — Template CRUD + back-compat
  re-exports so every existing import path keeps resolving.
- lease_template_placeholder_service.py (138 LOC) — AI-driven placeholder
  suggestion orchestration. Also owns TemplateNotFoundError and
  extract_text_from_upload — the shared seam that breaks the potential
  circular import between CRUD and render modules.
- lease_template_render_service.py (45 LOC) — load_template_source_texts,
  the helper that fetches raw template file bytes for the signed-lease PDF
  pipeline.

Circular import analysis: lease_template_placeholder_service has no imports
from lease_template_service; lease_template_render_service imports only
TemplateNotFoundError from the placeholder module. No cycles.

Class identity: TemplateNotFoundError is defined once (placeholder module)
and re-exported via lease_template_service. All existing consumers that
import it from the original module get the same class object — catch blocks
continue to work without change. Verified with assert TNF is TNF2.

No behavior change; pure structural split. All existing imports preserved
via re-exports in lease_template_service. 25/25 tests pass.
